refactor: export shared functions
