+++ app/src/components/others/CafeCarousel/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { CafeInfo } from '../../../utils/type';
 import CarouselMainImage from '../CarouselMainImage';
 import './index.css';
-import CarouselCol from '../CarouselCol';
+import CarouselVertical from '../CarouselVertical';
 import { StyledCarouselImage } from '../../../utils/styled';
 
 type CafeImageCarouselProps = {
@@ -13,15 +13,14 @@ type CafeImageCarouselProps = {
 const CafeCarousel = ({cafes, setCafe}: CafeImageCarouselProps) => {
     return(
         <div className="carousel-container">
-            <CarouselCol title="Carousel">
+            <CarouselVertical title="Carousel">
             {cafes?.map((cafe, index) => {
                 return(
                     <StyledCarouselImage key={index}>
                         <CarouselMainImage cafe={cafe} setCafe={setCafe}/>
                     </StyledCarouselImage>
             )})}
-            </CarouselCol>
-       
+            </CarouselVertical>
         </div>
     )
 }
