+++ app/src/components/others/CafeDetailImageCarousel/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyledCarouselImage } from '../../../utils/styled';
 import { CafeInfo } from '../../../utils/type';
-import CarouselRow from '../CarouselRow';
+import CarouselHorizontal from '../CarouselHorizontal';
 import CarouselDetailImage from '../CarouselDetailImage';
 import './index.css';
 
@@ -12,7 +12,7 @@ type CafeDetailImageCarouselProps = {
 const CafeDetailImageCarousel = ({cafe}: CafeDetailImageCarouselProps) => {
     return(
         <div className="detail-carousel-container">
-            <CarouselRow title="Carousel">
+            <CarouselHorizontal title="Carousel">
             {cafe.imageUris.map((image, index) => {   
                 return(
                     <StyledCarouselImage key={index}>
@@ -20,7 +20,7 @@ const CafeDetailImageCarousel = ({cafe}: CafeDetailImageCarouselProps) => {
                     </StyledCarouselImage>
                 )
             })}
-            </CarouselRow>
+            </CarouselHorizontal>
         </div>
     )
 }
