+++ app/src/components/others/CarouselVertical/index.css
