+++ app/src/utils/constant.tsx
@@ -7,3 +7,5 @@ export const UP = "UP";
 export const DOWN = "DOWN";
 export const LEFT = "LEFT";
 export const RIGHT = "RIGHT";
+
+export const initialCarouselState = { pos: 0, sliding: false, dir: null };
